requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
typing-extensions>=4.0.0
streamlit>=1.32.0
python-docx>=0.8.11
graphviz>=0.20.1 
sentence-transformers>=2.2.2
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0